removing 8 interface limit restructuring interface IDs and configuration
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@28312 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/impl/com/untangle/jnetcap/Netcap.java
@@ -29,7 +29,8 @@ import java.util.Collections;
 import java.net.InetAddress;
 import org.apache.log4j.Logger;
 
-public final class Netcap {
+public final class Netcap
+{
     public static final short IPPROTO_UDP  = 17;
     public static final short IPPROTO_TCP  = 6;
     public static final short IPPROTO_ICMP = 1;
@@ -221,7 +222,7 @@ public final class Netcap {
     /**
      * Configure the netcap interface array
      */
-    public void configureInterfaceArray( byte intfIndexArray[], String interfaceArray[] )
+    public void configureInterfaceArray( int intfIndexArray[], String interfaceArray[] )
         throws JNetcapException
     {
         try {
@@ -397,5 +398,5 @@ public final class Netcap {
     /**
      * Function to configure the netcap interface array 
      */
-    private native void cConfigureInterfaceArray( byte intfIndexArray[], String interfaceArray[] );
+    private native void cConfigureInterfaceArray( int intfIndexArray[], String interfaceArray[] );
 }
