+++ uvm-lib/api/com/untangle/uvm/node/firewall/intf/IntfSingleMatcher.java
@@ -1,36 +1,4 @@
-/*
- * $HeadURL$
- * Copyright (c) 2003-2007 Untangle, Inc.
- *
- * This library is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License, version 2,
- * as published by the Free Software Foundation.
- *
- * This library is distributed in the hope that it will be useful, but
- * AS-IS and WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE, TITLE, or
- * NONINFRINGEMENT.  See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Linking this library statically or dynamically with other modules is
- * making a combined work based on this library.  Thus, the terms and
- * conditions of the GNU General Public License cover the whole combination.
- *
- * As a special exception, the copyright holders of this library give you
- * permission to link this library with independent modules to produce an
- * executable, regardless of the license terms of these independent modules,
- * and to copy and distribute the resulting executable under terms of your
- * choice, provided that you also meet, for each linked independent module,
- * the terms and conditions of the license of that module.  An independent
- * module is a module which is not derived from or based on this library.
- * If you modify this library, you may extend this exception to your version
- * of the library, but you are not obligated to do so.  If you do not wish
- * to do so, delete this exception statement from your version.
- */
-
+/* $HeadURL$ */
 package com.untangle.uvm.node.firewall.intf;
 
 import java.util.HashMap;
@@ -64,12 +32,12 @@ public final class IntfSingleMatcher extends IntfDBMatcher
     private static final IntfDBMatcher VPN_MATCHER;
 
     /* Map of interfaces to their matcher */
-    static Map<Byte,IntfSingleMatcher> CACHE = new HashMap<Byte,IntfSingleMatcher>();
+    static Map<Integer,IntfSingleMatcher> CACHE = new HashMap<Integer,IntfSingleMatcher>();
 
     /* The interface this matcher matches */
-    private final byte intf;
+    private final int intf;
 
-    private IntfSingleMatcher(byte intf)
+    private IntfSingleMatcher(int intf)
     {
         this.intf = intf;
     }
@@ -80,7 +48,7 @@ public final class IntfSingleMatcher extends IntfDBMatcher
      * @param intf Interface to test.
      * @return True if the interface matches.
      */
-    public boolean isMatch(byte iface, byte otherIface)
+    public boolean isMatch(int iface, int otherIface)
     {
         /* These interfaces always match true */
         if (IntfConstants.UNKNOWN_INTF == iface
@@ -95,7 +63,7 @@ public final class IntfSingleMatcher extends IntfDBMatcher
 
     public String toDatabaseString()
     {
-        return Byte.toString(intf);
+        return Integer.toString(intf);
     }
 
     public String toString()
@@ -131,7 +99,7 @@ public final class IntfSingleMatcher extends IntfDBMatcher
         return VPN_MATCHER;
     }
 
-    public static IntfDBMatcher makeInstance(byte intf)
+    public static IntfDBMatcher makeInstance(int intf)
     {
         IntfSingleMatcher cache = CACHE.get(intf);
